InputFieldPropsとButtonPropsを各コンポーネント内に定義するようリファクタリング Firebaseエラー処理をlib配下に切り出し、Propsの責務を整理
