+++ backend/app/controllers/users_controller.rb
@@ -17,6 +17,14 @@ class UsersController < ApplicationController
     @user = current_user
   end
 
+  def me
+    if current_user
+      render json: current_user.as_json(only: [:id, :name, :email, :uid])
+    else
+      head :not_found
+    end
+  end
+
   private
 
   def user_params
