+++ frontend-react/src/components/CustomButton.tsx
@@ -1,0 +1,20 @@
+interface ButtonProps {
+  onClick?: () => void
+  children: React.ReactNode
+  icon?: string
+  className?: string
+}
+
+const Button = ({ onClick, children, icon, className = "" }: ButtonProps) => {
+  return (
+    <button
+      onClick={onClick}
+      className={`md:w-3/4 w-80 my-4 text-blue-600 border-4 border-blue-400 border-double px-3 py-2 ${className}`}
+    >
+      {icon && <span className={`${icon} w-5 h-5 float-left`}></span>}
+      <span className="px-3">{children}</span>
+    </button>
+  )
+}
+
+export default Button
