+++ frontend-react/src/pages/SignupPage.tsx
@@ -1,4 +1,5 @@
-import { ButtonProps, RailsApiError } from "@/types"
+import Button from "@/components/CustomButton"
+import { RailsApiError } from "@/types"
 import { useState } from "react"
 import { auth } from "@/lib/firebase"
 import { GoogleAuthProvider, signInWithPopup, User } from "firebase/auth"
@@ -6,30 +7,7 @@ import { useApiClient } from "@/hooks/useApiClient"
 import { useSetAuth } from "@/context/useAuthContext"
 import { FirebaseError } from "firebase/app"
 import { useNavigate } from "react-router-dom"
-
-const getErrorMessage = (error: unknown): string => {
-  if (error instanceof FirebaseError) {
-    switch (error.code) {
-      case "auth/account-exists-with-different-credential":
-        return "このメールアドレスは別の認証方法で登録されています。元の認証方法でログインしてください。"
-      case "auth/popup-closed-by-user":
-        return "ポップアップが閉じられました。もう一度試してください。"
-      case "auth/cancelled-popup-request":
-        return "複数の認証リクエストが行われました。しばらくしてから再試行してください。"
-      case "auth/popup-blocked":
-        return "ポップアップがブロックされました。ブラウザの設定を確認してください。"
-      case "auth/network-request-failed":
-      case "auth/internal-error":
-        return "ネットワーク接続に問題があります。接続を確認して再試行してください。"
-      case "auth/credential-already-in-use":
-        return "このGoogleアカウントはすでに別のFirebaseアカウントにリンクされています。"
-      default:
-        return "エラーが発生しました。しばらくしてから再試行してください。"
-    }
-  }
-
-  return "予期しないエラーが発生しました。"
-}
+import { getFirebaseErrorMessage } from "@/lib/firebaseErrorHandler"
 
 const SignupPage = () => {
   const [error, setError] = useState<string | null>(null)
@@ -67,7 +45,7 @@ const SignupPage = () => {
       console.log("home画面は次のissueで作成予定！")
     } catch (error: unknown) {
       if (error instanceof FirebaseError) {
-        setError(getErrorMessage(error))
+        setError(getFirebaseErrorMessage(error))
         return
       }  
 
@@ -118,18 +96,6 @@ const SignupPage = () => {
         </div>
       </div>
     </div>
-  )
-}
-
-const Button = ({ onClick, children, icon, className = "" }: ButtonProps) => {
-  return (
-    <button
-      onClick={onClick}
-      className={`md:w-3/4 w-80 my-7 text-blue-600 border-4 border-blue-400 border-double px-3 py-2 ${className}`}
-    >
-      {icon && <span className={`${icon} w-5 h-5 float-left`}></span>}
-      <span className="px-3">{children}</span>
-    </button>
   )
 }
